==> 09_model_selection_loo.Rmd <==
---
title: "Alternative Models"
---

## Alternative Models

```{r}
## library here package 
library(here)
## Load helper functions and library all packages used for analysis 
source(here("code", "models", "mrp", "helpers.R"))
```

## read in data for model 

```{r}
## prepared for model  
df_formodel <- read_rds(here('data', 'mrp', 'input', 'df_formodel_waves1_6.rds')) 

## top code total number of reported contacts at 30 
top_code <- function(x) {
  x <- case_when(
    x > 30 ~ 30,
    TRUE ~ x
  )
}

## topcode data.frame 
df_formodel_topcode <- df_formodel %>%
  mutate(across(c(starts_with("alter_age")), top_code))
```

Alternative model specifications with different time trends. 

```{r}
## model with national-level time trend polynomial 
## Total execution time: 2052.9 seconds.
mobility_model0a <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
    poly(day_std, 3), 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4,
  backend = "cmdstanr",
  threads = threading(4),
  cores = 16,
  iter = 4000)

## write out model 
write_rds(mobility_model0a, file = here('data', 'mrp', 'models', 'mobility_model0a.rds'))

## model with state-level hierarchical cubic time trend 
## Total execution time: 2065.0 seconds.
mobility_model0b <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
  (poly(day_std, 3) | state), ## state level pooled polynomial trend 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4,
  backend = "cmdstanr",
  threads = threading(4),
  cores = 16,
  iter = 4000)

## write out model 
write_rds(mobility_model0b, file = here('data', 'mrp', 'models', 'mobility_model0b.rds'))

## Model with national-level polynomial and state-level hierarchical cubic time trend  
## Total execution time: 2065.0 seconds.
mobility_model0c <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
   (poly(day_std, 3) | state) +  ## state level pooled polynomial trend 
    poly(day_std, 3), 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4,
  backend = "cmdstanr",
  threads = threading(4),
  cores = 16,
  iter = 4000)

## write out model 
write_rds(mobility_model0c, file = here('data', 'mrp', 'models', 'mobility_model0c.rds'))

## mobility model 1
mobility_model1 <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~ 
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
    (1 | state) + 
    rolling_pca_3wks, 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4,
  backend = "cmdstanr",
  threads = threading(4),
  cores = 16,
  iter = 4000)

## write out model 
write_rds(mobility_model1, file = here('data', 'mrp', 'models', 'mobility_model1.rds'))

## pca + weekly monthly trends 
## Total execution time: 5696.4 seconds.
mobility_model2 <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~ 
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
    rolling_pca_3wks + 
    (poly(day_std, 3) | state), 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4,
  cores = 16,
  backend = "cmdstanr",
  threads = threading(4),
#  control = list(adapt_delta = 0.9), 
  iter = 4000)


## write out model 
write_rds(mobility_model2, file = here('data', 'mrp', 'models', 'mobility_model2.rds'))

## mobility + monthly national cubic + state-level hierarchical term 
## Total execution time: 5346.0 seconds.
mobility_model3 <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~ 
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
    rolling_pca_3wks + ## mobility data 
    poly(day_std, 3) + ## national polynomial trend 
    (poly(day_std, 3) | state), ## state level pooled polynomial trend 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4,
  cores = 16,
  backend = "cmdstanr",
  threads = threading(4),
#  control = list(adapt_delta = 0.9), 
  iter = 4000)

## write out model 
write_rds(mobility_model3, file = here('data', 'mrp', 'models', 'mobility_model3.rds'))

## Total execution time: 5346.0 seconds.
mobility_model4 <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~ 
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
    (1 | state ) + 
    rolling_pca_3wks + ## mobility data 
    poly(day_std, 3), ## national polynomial trend 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4,
  cores = 16,
  backend = "cmdstanr",
  threads = threading(4),
#  control = list(adapt_delta = 0.9), 
  iter = 4000)


## write out model 
write_rds(mobility_model4, file = here('data', 'mrp', 'models', 'mobility_model4.rds'))

```

Post-stratification code is in the `03_poststratification_daily.Rmd` Notebook. 


```{r}
## prepared for model  
df_formodel <- read_rds(here('data', 'mrp', 'input', 'df_formodel_waves1_6.rds')) 

## top code total number of reported contacts at 30 
top_code <- function(x) {
  x <- case_when(
    x > 100 ~ 100,
    TRUE ~ x
  )
}

## topcode data.frame 
df_formodel_topcode_100 <- df_formodel %>%
  mutate(across(c(starts_with("alter_age")), top_code))


## add on google mobility data 
google_pca_data <- read_rds(file=here('data', 'mrp', 'input', 'google_mobility_pca.rds'))

df_formodel_topcode_100 <- df_formodel_topcode_100 %>% 
  mutate(date = as_date(startdate)) %>% 
  left_join(google_pca_data %>% rename(state = state), by = c("state", "date"))
```


```{r}
## mobility + monthly national cubic + state-level hierarchical cubic term 
## Total execution time: 5346.0 seconds.
final_mobility_model_topcode_100 <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~ 
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
    agecat:gender + 
    rolling_pca_3wks + ## mobility data 
    poly(day_std, 3) + ## national polynomial trend 
    (poly(day_std, 3) | state), ## state level pooled polynomial trend 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode_100,
  chains = 4,
  cores = 16,
  backend = "cmdstanr",
  threads = threading(4),
  #  control = list(adapt_delta = 0.9), 
  iter = 4000)

```

```{r}
coef_nonhh_model %>% 
  inner_join(coef_nonhh_model_topcode100, by = "rowname") %>% 
  mutate(diff = 100 * abs(Estimate.x - Estimate.y) / ((Estimate.x + Estimate.y) / 2)) %>% 
  View()

comparison_plot <- coef_nonhh_model %>% 
  mutate(model = "topcode 30") %>% 
  bind_rows(coef_nonhh_model_topcode100 %>% 
              mutate(model = "topcode 100")) %>% 
 # mutate(diff = 100 * abs(Estimate.x - Estimate.y) / ((Estimate.x + Estimate.y) / 2)) %>% 
  ggplot(aes(y = rowname, x = Estimate, xmin = Q2.5, xmax = Q97.5, color = model)) + 
  geom_pointrange(position = position_dodge(width = 1)) + 
  theme_cowplot() + 
  ggsci::scale_color_lancet() 

ggsave(plot = comparison_plot, filename = here("code", "models", "mrp", "figures", "model_comparison_topcode.png" ), height = 30, width = 10)
```



```{r}
## write out model
mobility_model1 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model1.rds'))
mobility_model2 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model2.rds'))
mobility_model3 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model3.rds'))
mobility_model4 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model4.rds'))
mobility_model5 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model5.rds'))
mobility_model6 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model6.rds'))
mobility_model7 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model7.rds'))
mobility_model8 <- read_rds(file = here('data', 'mrp', 'models', 'mobility_model8.rds'))
```


```{r}
mobility_model1_loo <- loo(mobility_model1,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model2_loo <- loo(mobility_model2,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model3_loo <- loo(mobility_model3,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model3.1_loo <- loo(mobility_model3.1,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model4_loo <- loo(mobility_model4,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model5_loo <- loo(mobility_model5,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model6_loo <- loo(mobility_model6,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model7_loo <- loo(mobility_model7,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model8_loo <- loo(mobility_model8,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model0a_loo <- loo(mobility_model0a,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model0b_loo <- loo(mobility_model0b,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model0c_loo <- loo(mobility_model0c,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

mobility_model0d_loo <- loo(mobility_model0d,
                           nsamples = 1000,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))

hh_model_loo <- loo(hh_model,
                           nsamples = 100,
                           draws = NULL,
                           cores = getOption("mc.cores", 1))
```

```{r}
loo_compare(mobility_model0a_loo, mobility_model0b_loo, mobility_model0c_loo, mobility_model0d_loo, mobility_model1_loo, mobility_model2_loo, mobility_model3_loo, mobility_model4_loo, mobility_model5_loo, mobility_model6_loo, mobility_model7_loo, mobility_model8_loo)
```
```{r}
loo_compare(mobility_model0c_loo, mobility_model3_loo, mobility_model3.1_loo)
```
